Use more robust check before using noConflict() mode
Basically detect the case when L is already the same Leaflet library instance
that was loaded by some other extension, in which case there is no need to use
noConflict. This does not address the problem of plugins assuming single global
L when there are several different Leaflet libs, but it does address #337, since
other extension share the same Leaflet instance.