Allow rendering GeoJSON points with CircleMarkers.
Before this commit GeoJSON points where always rendered as 'map pins'.
Now users can pass all CircleMarker options to `GeoJSON` using the
`point_style` argument. If the argument is present and non-empty,
points will be rendered using CircleMarkers. If the argument is
ommitted or empty, points are rendered as 'map pins'.

This could serve as a solution for #188 and #362.