Add doc section on basic layout for Map #750
How to customise the layout is not obvious to new users
not already familiar with ipywidgets. Screen sizes these days
are such the default layout appears very horizontal and
users appear often inclined to change the default height. The
first experience to find out how can be smoothed a bit
via documentation.